# docs/reference/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Thoth API",
    "version": "0.1.0",
    "description": "Live OSINT layers, intelligence, watches and on-demand lookups. See docs/reference/api.md for conventions (authentication, list and error shapes, the event object).",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:4000",
      "description": "API, direct (host only)"
    },
    {
      "url": "/",
      "description": "Through the app (access token when gated)"
    }
  ],
...
  "paths": {
...
    "/api/openapi.json": {
      "get": {
        "operationId": "getOpenapiJson",
        "tags": [
          "System"
        ],
        "summary": "This OpenAPI document",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
...
    "/api/layers/{layer}": {
      "get": {
        "operationId": "getLayersLayer",
        "tags": [
          "Layers"
        ],
        "summary": "Layer slice",
        "description": "Without `z`: the newest 500 events. With `z`: a viewport slice with spatial sampling, plus matched, limit and truncated.",
        "parameters": [
          {
            "name": "layer",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "z",
            "in": "query",
            "required": false,
            "description": "Map zoom. With it, the response is a viewport slice.",
            "schema": {
              "type": "number",
              "minimum": 0,
              "maximum": 24
            }
          },
          {
            "name": "bbox",
            "in": "query",
            "required": false,
            "description": "Viewport `w,s,e,n` in degrees; w > e crosses the antimeridian.",
            "schema": {
              "type": "string",
              "maxLength": 120
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "ok": {
            "const": false
          },
          "error": {
            "type": "string"
          }
        },
        "required": [
          "ok",
          "error"
        ]
      }
    },
    "responses": {
...
      "RateLimited": {
        "description": "Rate limited: retry after `Retry-After` seconds",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "headers": {
          "Retry-After": {
            "schema": {
              "type": "integer"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerWriteKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "API_WRITE_KEY, for POST and DELETE"
      },
      "headerWriteKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Thoth-Key"
      }
    }
  }
}